% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_and_install_package.R
\name{check_and_install_package}
\alias{check_and_install_package}
\title{Check and install package}
\usage{
check_and_install_package(packageName)
}
\arguments{
\item{packageName}{Name of the package}
}
\description{
This function checks whether a package has been installed or not. If it hasn't it will install it.
}
\examples{
check_and_install_package()
}
