% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_all_dependendices.R
\name{py_version_options}
\alias{py_version_options}
\title{List all python versions}
\usage{
py_version_options(regex = ".*", ...)
}
\arguments{
\item{regex}{regular expression to use on path}

\item{...}{additional arguments to py_discover_config}
}
\value{
character vector of python paths that are available and match regex
}
\description{
List all python versions
}
