% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_cam.R
\name{visualize_cam}
\alias{visualize_cam}
\title{visualize_cam}
\usage{
visualize_cam(model, img, save_folder = "Class Activation Map")
}
\arguments{
\item{model}{Keras NN object}

\item{img}{The image of the file you want to test}

\item{save_folder}{The name of the folder you want to save the images in. "Saliency_Map" is the default.}
}
\description{
To do
}
\examples{
visualize_cam()
}
